multiplication algorithm should now be working

+++ src/C++/main.cpp
@@ -4,8 +4,8 @@
 int main()
 {
 	int a[]{ 1,2,3,4 };
-	matrix<int>* m = new serial_matrix<int>(2000, 2000);
-	matrix<int>* n = new serial_matrix<int>(2000, 2000);
+	matrix<float>* m = new serial_matrix<float>(2000, 2000);
+	matrix<float>* n = new serial_matrix<float>(2000, 2000);
 	auto b = ((*m) * (*n));
 	int val = b->get_val(1, 1);
 	printf("%d\n", val);
